Create tutte le repositories di Utenti, Sorgenti e Aziende

+++ src/main/java/it/smartwater/be/models/aziende/Manutenzione.java
@@ -33,7 +33,7 @@ public class Manutenzione {
 
     @Column(name="idTipoManutenzione")
     @NotNull
-    private long idTipoManutenzione;
+    private int idTipoManutenzione;
 
 
     @OneToOne
